requirements.txt
python-telegram-bot == 21.4
python-telegram-bot[job-queue]
Pillow == 9.5.0
python-dotenv
opencv-python